+++ keel-codec-http/src/commonTest/kotlin/io/github/fukusaka/keel/codec/http/CompressionHandlerTest.kt
@@ -963,6 +963,8 @@ class CompressionHandlerTest {
         override fun propagateActive() {}
         override fun propagateInactive() {}
         override fun propagateReadComplete() {}
+
+        override fun propagateFlushComplete() {}
         override fun propagateError(cause: Throwable) {}
         override fun propagateUserEvent(event: Any) {}
         override fun propagateWritabilityChanged(isWritable: Boolean) {}