+++ keel-codec-http/src/commonTest/kotlin/io/github/fukusaka/keel/codec/http/HttpRequestDecompressionFixture.kt
@@ -217,6 +217,8 @@ internal abstract class HttpRequestDecompressionFixture {
         override fun propagateActive() {}
         override fun propagateInactive() {}
         override fun propagateReadComplete() {}
+
+        override fun propagateFlushComplete() {}
         override fun propagateError(cause: Throwable) {}
         override fun propagateUserEvent(event: Any) {}
         override fun propagateWritabilityChanged(isWritable: Boolean) {}